add actors runner
